remove some extra line  from the space
Gatsby continues to attract popular and scholarly attention. Scholars emphasize the novel's treatment of social class, inherited versus self-made wealth, gender, race, and environmentalism, and its cynical attitude towards the American Dream. The Great Gatsby is widely considered to be a literary masterpiece and a contender for the title of the Great American Novel.